+++ src/components/CucmRoutingPage/index.js
@@ -1,6 +1,7 @@
 /**@jsx jsx*/
 import {jsx, css} from '@emotion/core'
 import CucmList from "./components/CucmList";
+import CucmRoutesTable from "./components/CucmRoutesTable";
 import {Button, Input} from 'reactstrap'
 import {useState} from 'react'
 
@@ -8,14 +9,14 @@ const CucmRoutingPage = (props) => {
   const [pattern, setPattern] = useState('')
   const [cucm, setCucm] = useState('')
   const [loading, setLoading] = useState(false)
-  const [extFilter, setExtFilter] = useState({cucm: '', num: ''})
-  const onClickHandler = () => cucm && pattern && setExtFilter(prevState => ({cucm, num: pattern}))
+  const [extFilter, setExtFilter] = useState({cucm: '', pattern: ''})
+
+  const onClickHandler = () => cucm && pattern && setExtFilter(prevState => ({cucm, pattern}))
   const onChangePattern = (e) => setPattern(e.target.value)
   const onChangeSelectedCucm = ({accessor, value}) => {
     console.log('cucm', value[0], accessor)
     setCucm(value.pop())
   }
-  console.log({cucm, pattern, test: !!(cucm && pattern)})
 
   return (
     <div css={css`display: flex; flex-direction: column; height: 100%`}>
@@ -27,10 +28,12 @@ const CucmRoutingPage = (props) => {
           <div className="ml-2 ">
             <Input onChange={onChangePattern} />
           </div>
-          <Button className="ml-2" css={css`background-color: rgb(62,148,224)`} onClick={onClickHandler} disabled={!(cucm && pattern)}>Search</Button>
+          <Button className="ml-2" css={css`background-color: rgb(62,148,224)`} onClick={onClickHandler} disabled={!(cucm && pattern) || loading}>Search</Button>
         </div>
       </div>
-      <div css={css`flex-grow: 1; min-height: 0; background-color: #005cbf`}>Table</div>
+      <div css={css`flex-grow: 1; min-height: 0`}>
+        <CucmRoutesTable extFilter={extFilter} setLoading={setLoading} />
+      </div>
     </div>
   )
 }