+++ src/constants/urls.js
@@ -13,7 +13,7 @@ console.log("BASE API URL", BASE_URL)
 //internal links
 export const DICT_REG_CENTERS_MAPPING= `/vra/rc/mapping`
 export const UNREGISTERED_PHONES = `/vra/tools/unregisteredPhones`
-export const CUCM_ROUTING = '/vra/tools/cucmRouting'
+export const CUCM_ROUTES = '/vra/tools/cucmRouting'
 export const TEST_TOOLS = `/vra/tools/testing`
 //external links
 export const LOCATIONS = `${BASE_URL}/locations`
@@ -39,5 +39,6 @@ export const DICT_PHONE_LOGS= `${BASE_URL}/log/phone`
 // for unregistered phones in office
 export const OFFICE_LIST_URL = `${BASE_URL}/phone/officeList.json`
 //for cucm routing page
-export const CUCM_LIST = 'http://10.99.120.223/cucm_list/'
+export const GET_CUCM_LIST = 'http://10.99.120.223:8000/cucm_list/'
+export const GET_CUCM_ROUTES = 'http://10.99.120.223:8000/outgoing_routes/'